At least it doesn't use nocheck
